+++ src/hooks/index.js
@@ -1,4 +1,5 @@
 export * from "./useBoolean";
 export { useClickAway } from "./useClickAway";
+export { useLockBodyScroll } from "./useLockBodyScroll";
 export { useMount } from "./useMount";
 export { useMountLayout } from "./useMountLayout";
